backend/requirements.txt
langchain
langchain-community
langchain-google-genai
pypdf
sentence-transformers
google-generativeai
tiktoken
faiss-cpu
unstructured
tokenizers
openai
flask
flask-cors
pymongo
python-dotenv
nltk
requests

# Image handling dependencies
Pillow>=9.0.0
gridfs>=0.5.0
Werkzeug>=2.0.0
